JG-2007-06-19-0: Bugfix:    Grid Engine commands failed, when the primary group                             of the user was too big in groups file/map.                  Issue:     2249                  Bugtraq:   6568575                  Review:    RD, AA Issue number: Obtained from: Submitted by: Reviewed by:

+++ source/daemons/qmaster/sge_c_gdi.c
@@ -293,7 +293,7 @@ sge_c_gdi(char *host, sge_gdi_request *request, sge_gdi_request *response,
    if (!sge_security_verify_user(request->host, request->commproc, 
                                  request->id, user)) {
       CRITICAL((SGE_EVENT, MSG_SEC_CRED_SSSI, user, request->host, 
-                request->commproc, request->id));
+                request->commproc, (int)request->id));
       answer_list_add(&(response->alp), SGE_EVENT, 
                       STATUS_ENOSUCHUSER, ANSWER_QUALITY_ERROR);
       DEXIT;
